backend/requirements.txt
Flask==3.1.0
Flask_Cors==3.0.10
flask_sqlalchemy==3.1.1
python_docx==1.1.2
Requests==2.32.3
Werkzeug==3.1.3
XMind==1.2.0
